+++ src/selfDefined/EthSignRequest.ts
@@ -1,8 +1,8 @@
-import { CryptoKeypath } from './CryptoKeypath';
-import { decodeToDataItem } from './lib/cbor-sync';
-import { DataItem } from './lib/DataItem';
-import { RegistryItem } from './RegistryItem';
-import { RegistryTypes } from './RegistryType';
+import { CryptoKeypath } from '../CryptoKeypath';
+import { decodeToDataItem } from '../lib/cbor-sync';
+import { DataItem } from '../lib/DataItem';
+import { RegistryItem } from '../RegistryItem';
+import { RegistryTypes } from '../RegistryType';
 
 enum Keys {
     requestId = 1,
