+++ src/selfDefined/index.ts
@@ -1,0 +1,2 @@
+export { EthSignRequest, DataType } from './EthSignRequest';
+export { ETHSignature } from './EthSignature';
